+++ lib/src/index.ts
@@ -2,7 +2,11 @@ import { resolvePageComponent } from 'laravel-vite-plugin/inertia-helpers'
 import type { DefineComponent } from 'vue'
 import { VueTypescriptable } from './vue/plugin'
 
-async function resolve(name: string, glob: Record<string, () => Promise<unknown>>): Promise<DefineComponent> {
+function resolveTitle(title: string, appName: string, seperator = '·'): string {
+  return title ? `${title} ${seperator} ${appName}` : `${appName}`
+}
+
+async function resolvePages(name: string, glob: Record<string, () => Promise<unknown>>): Promise<DefineComponent> {
   return resolvePageComponent(`./Pages/${name}.vue`, glob) as Promise<DefineComponent>
 }
 
@@ -28,5 +32,6 @@ export {
 
 export {
   VueTypescriptable,
-  resolve,
+  resolveTitle,
+  resolvePages,
 }
