+++ lib/tsconfig.json
@@ -1,22 +1,22 @@
 {
   "compilerOptions": {
+    "target": "ESNext",
+    "jsx": "preserve",
+    "jsxImportSource": "vue",
+    "lib": ["ESNext", "DOM"],
+    "module": "ES2020",
+    "moduleResolution": "Node",
     "paths": {
       "@": ["./src"],
       "@/*": ["./src/*"],
       "~": ["./"],
       "~/*": ["./*"]
     },
-    "target": "ESNext",
-    "jsx": "preserve",
-    "lib": ["ESNext", "DOM"],
-    "module": "ES2020",
-    "moduleResolution": "Node",
-    "noImplicitAny": false,
     "strict": true,
     "strictNullChecks": true,
-    "esModuleInterop": true,
+    "noImplicitAny": false,
     "declaration": true,
-    "jsxImportSource": "vue"
+    "esModuleInterop": true
   },
   "include": [
     "src",
